Real-world integrated benchmark of wasmer2 vs wasmer0
In this benchmark, we compare wasmer2 vs wasmer0 on a real-world
contract. First, run

  export RUSTFLAGS='-C link-arg=-s'
  export CARGO_PROFILE_RELEASE_LTO='full'
  export CARGO_PROFILE_RELEASE_CODEGEN_UNITS='1'

  cargo build \
    --target wasm32-unknown-unknown \
    --release \
    -p aurora-engine \
    --no-default-features \
    --features=contract \
    -Z avoid-dev-deps

To build the contract

Then, run

  cargo test --package aurora-engine-tests --lib --features mainnet-test --release \
    -- tests::uniswap::test_uniswap_input_multihop --exact --nocapture

to run the actual benchmark with wasmer2.

Note the *last* `Wasmer2   :` line. On my machine, it is

Wasmer2   : 26.704477ms

Change to Wasmer0 by applying the following diff

  diff --git a/engine-tests/src/test_utils/mod.rs b/engine-tests/src/test_utils/mod.rs
  index c512ad4..fb944d3 100644
  --- a/engine-tests/src/test_utils/mod.rs
  +++ b/engine-tests/src/test_utils/mod.rs
  @@ -186,7 +186,7 @@ impl AuroraRunner {

           let mut config = self.wasm_config.clone();
           config.regular_op_cost = 0;
  -        let vm_kind = VMKind::Wasmer2;
  +        let vm_kind = VMKind::Wasmer0;
           let t = std::time::Instant::now();
           let (maybe_outcome, maybe_error) = vm_kind.runtime(config).unwrap().run(
               &self.code,

Re-running the benchmark, I get

Wasmer0   : 20.289034ms

Note that the benchmark is run with regular_op_cost set to 0 (that is,
without gas instrumentation).

Note that for micro-benchmarks without host function calls I anecdotally
observed as much as 2X difference between wasmer2 and wasmer0 and I
think I've also noticed that host function calls are much faster in
wasmer2. That is, what we might be observing here is not 20% difference
overall, but 50% difference in pure wasm execution.