correlate eth and near gas
To run the benchmark

```
$ make mainnet-test.wasm
$ cargo test --release --package aurora-engine-tests --lib --features mainnet-test \
  -- tests::sanity::test_solidity_pure_bench --exact --nocapture
```

Results on my machine:

```
wasm_time = 399.154259ms
real_time = 46.559449ms
eth  gas = 10_323_995
near gas = 1_288_932_851_101_824
```

That is:

10 million eth gas (size of the block on etherium)
1288 near terra gas
400ms wall-clock time

All three numbers seem consistent, using

25eth gas  = 1us
1near tgas = 1ms

and the 3x safety multiplier.

That is, we expect 10**7 eth gas to take 0.4 seconds of computation, and
exactly that happens. The cost in neargas corresponds to 1.2 seconds,
due to the safety multiplier.

Note that in the benchmark we reduce the size of the in-memory buffer --
memory is costly in evm, and having a large buffer skews the benchmark